init 8259A or init idt has problem!

+++ kernel/start.c
@@ -23,11 +23,11 @@ PUBLIC void cstart()
     *p_gdt_limit = GDT_SIZE * sizeof(DESCRIPTOR) - 1;
     *p_gdt_base  = (u32)&gdt;
 
-    /* idt_ptr[6] 共 6 个字节：0~15:Limit  16~47:Base。用作 sidt/lidt 的参数。*/
+    /* idt_ptr[10] 共 10 个字节：0~15:Limit  16~47:Base。用作 sidt/lidt 的参数。*/
     u16* p_idt_limit = (u16*)(&idt_ptr[0]);
-    u32* p_idt_base  = (u32*)(&idt_ptr[2]);
+    u64* p_idt_base  = (u64*)(&idt_ptr[2]);
     *p_idt_limit = IDT_SIZE * sizeof(GATE) - 1;
-    *p_idt_base  = (u32)&idt;
+    *p_idt_base  = (u64)&idt;
 
     init_prot();
